Appeals v2 breadcrumbs (#6801)
* Don't use AppealLayout

* Add the breadcrumb

* Add unit tests

* Update breadcrumb content

* Go back to a functional component
