+++ test/claims-status/components/appeals-v2/AppealInfo.unit.spec.jsx
@@ -2,18 +2,26 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import { expect } from 'chai';
 import { merge } from 'lodash';
-import AppealInfo from '../../../../src/js/claims-status/containers/AppealInfo';
+import { AppealInfo } from '../../../../src/js/claims-status/containers/AppealInfo';
+import { mockData } from '../../../../src/js/claims-status/utils/helpers';
 
-const appealIdParam = '7387389';
+const appealIdParam = mockData.data[0].id;
 
 const defaultProps = {
   params: { id: appealIdParam },
+  appeal: mockData.data[0]
 };
 
 describe('<AppealInfo/>', () => {
   it('should render', () => {
     const wrapper = shallow(<AppealInfo {...defaultProps}/>);
     expect(wrapper.type()).to.equal('div');
+  });
+
+  it('should render the breadcrumbs', () => {
+    const wrapper = shallow(<AppealInfo {...defaultProps}/>);
+    const breadcrumbs = wrapper.find('Breadcrumbs');
+    expect(breadcrumbs.length).to.equal(1);
   });
 
   it('should display a tabbed navigator', () => {
@@ -29,9 +37,17 @@ describe('<AppealInfo/>', () => {
     expect(wrapper.find('span.test').length).to.equal(1);
   });
 
+  it('should pass appeal as a prop to its children', () => {
+    const children = (<span className="test">Child Goes Here</span>);
+    const props = merge({}, { children }, defaultProps);
+    const wrapper = shallow(<AppealInfo {...props}/>);
+    expect(wrapper.find('span.test').prop('appeal')).to.eql(defaultProps.appeal);
+  });
+
   it('should have access to the appeal id in route params', () => {
     const wrapper = shallow(<AppealInfo {...defaultProps}/>);
     const appealId = wrapper.instance().props.params.id;
     expect(appealId).to.equal(appealIdParam);
   });
 });
+
